requirements.txt
asgiref==3.8.1
attrs==21.2.0
Django==5.2
djangorestframework==3.16.0
docopt==0.6.2
filelock==3.18.0
html5lib==1.1
idna==3.10
importlib-metadata==4.6.4
jsonschema==3.2.0
matplotlib==3.10.1
more-itertools==8.10.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pipenv==2025.0.1
psycopg2-binary==2.9.10
python-dotenv==0.19.2
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
ssh-import-id==5.11
supervisor==4.2.1
svglib==1.5.1
texttable==1.6.4
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.4.0
virtualenv==20.30.0
weasyprint==65.1
webencodings==0.5.1
websocket-client
xhtml2pdf==0.2.17
zipp==1.0.0
zopfli==0.2.3.post1
django-cors-headers
